ai-system/backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
openai
httpx
pydantic
requests
slowapi
